requirements.txt
langchain
langchain-community
langchain-openai
llama-index
openai
tiktoken
deeplake
cohere
wikipedia
ragas
transformers
torch
pandas
numpy
python-dotenv
